Settings: Add an option to force pre-O apps to use full screen aspect ratio
When an app target pre-O releases, the default max aspect ratio
is 1.86:1 which leads to ugly black areas on devices that have
screens with higher aspect ratio (for example Galaxy S8/S9).

This change adds an option to allow users to change aspect ratio
for pre-O apps to full screen aspect ratio.

Change-Id: I23ce8f9ebaee31e8add173d13cfe7d5a783c59ab
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>